pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "can_test"
version = "0.1.0"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "customtkinter==5.2.2",
    "darkdetect==0.8.0",
    "fastapi",
    "idna==3.10",
    "jinja2",
    "msgpack==1.0.8",
    "mypy==1.10.1",
    "mypy-extensions==1.0.0",
    "netifaces==0.11.0",
    "packaging==24.1",
    "pillow==11.0.0",
    "ply==3.11",
    "pyasn1==0.6.1",
    "pycryptodomex==3.21.0",
    "pyftdi==0.55.4",
    "pyserial==3.5",
    "pysnmp==7.1.12",
    "pysnmp-pyasn1==1.1.3",
    "pysnmp-pysmi==1.1.12",
    "pysnmplib==5.0.24",
    "python-can>=4.4.2",
    "python-multipart==0.0.17",
    "result",
    "typing-extensions",
    "uvicorn",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["can_test"]

[project.scripts]
can_test = "can_test.main:main"

[tool.uv]
package = true
